chore: opt slack-service
