+++ sample/slack-1.json
@@ -1,0 +1,9 @@
+{
+    "attachments": [
+        {
+            "text": "hello test",
+            "pretext": "pre-test",
+            "title": "hello title"
+        }
+    ]
+}
